+++ R/airsis_createMonitorObject.R
@@ -9,6 +9,9 @@
 #' @param zeroMinimum logical specifying whether to convert negative values to zero
 #' @param baseUrl base URL for data queries
 #' @param saveFile optional filename where raw CSV will be written
+#' @param existingMeta existing 'meta' dataframe from which to obtain metadata for known monitor deployments
+#' @param addGoogleMeta logicial specifying wheter to use Google elevation and reverse geocoding services
+#' @param addEsriMeta logicial specifying wheter to use ESRI elevation and reverse geocoding services
 #' @param ... additional parameters are passed to type-specific QC functions
 #' @return A \emph{ws_monitor} object with AIRSIS data.
 #' @description Obtains monitor data from an AIRSIS webservice and converts
@@ -64,6 +67,9 @@ airsis_createMonitorObject <- function(startdate=strftime(lubridate::now(),"%Y01
                                        zeroMinimum=TRUE,
                                        baseUrl="http://xxxx.airsis.com/vision/common/CSVExport.aspx?",
                                        saveFile=NULL,
+                                       existingMeta=NULL,
+                                       addGoogleMeta=FALSE,
+                                       addEsriMeta=FALSE,
                                        ...) {
 
   # Sanity checks
@@ -125,7 +131,10 @@ airsis_createMonitorObject <- function(startdate=strftime(lubridate::now(),"%Y01
   # NOTE:  This step will create a uniformly named set of properties and will
   # NOTE:  add site-specific information like timezone, elevation, address, etc.
   logger.debug("Creating 'meta' dataframe ...")
-  meta <- airsis_createMetaDataframe(tbl, provider, unitID, 'AIRSIS')
+  meta <- airsis_createMetaDataframe(tbl, provider, unitID, 'AIRSIS',
+                                     existingMeta = existingMeta,
+                                     addGoogleMeta = addGoogleMeta,
+                                     addEsriMeta = addEsriMeta)
   
   # Create 'data' dataframe of PM2.5 values organized as time-by-monitorID
   logger.debug("Creating 'data' dataframe ...")